+++ packages/feathers/src/declarations.ts
@@ -1,12 +1,18 @@
-import { EventEmitter } from 'events';
 import {
-  NextFunction, HookContext as BaseHookContext
-} from '@feathersjs/hooks';
+  EventEmitter, NextFunction, HookContext as BaseHookContext
+} from './dependencies';
 
 type SelfOrArray<S> = S | S[];
 type OptionalPick<T, K extends PropertyKey> = Pick<T, Extract<keyof T, K>>
 
-export { NextFunction };
+export type { NextFunction };
+
+export interface Paginated<T> {
+  total: number;
+  limit: number;
+  skip: number;
+  data: T[];
+}
 
 export interface ServiceOptions {
   events?: string[];
@@ -53,7 +59,7 @@ export type Service<T, D = Partial<T>> =
 export type ServiceInterface<T, D = Partial<T>> =
   Partial<ServiceMethods<T, D>>;
 
-export interface ServiceAddons<A, S> extends EventEmitter {
+export interface ServiceAddons<A = Application, S = Service<any, any>> extends EventEmitter {
   id?: string;
   hooks (options: HookOptions<A, S>): this;
 }
@@ -314,6 +320,8 @@ export interface HookContext<A = Application, S = any> extends BaseHookContext<S
 // Legacy hook typings
 export type LegacyHookFunction<A = Application, S = Service<any, any>> =
   (this: S, context: HookContext<A, S>) => (Promise<HookContext<Application, S> | void> | HookContext<Application, S> | void);
+
+export type Hook<A = Application, S = Service<any, any>> = LegacyHookFunction<A, S>;
 
 type LegacyHookMethodMap<A, S> =
   { [L in keyof S]?: SelfOrArray<LegacyHookFunction<A, S>>; } &