+++ packages/feathers/src/hooks/index.ts
@@ -1,5 +1,9 @@
-import { getManager, HookContextData, HookManager, HookMap, HOOKS, hooks, Middleware } from '@feathersjs/hooks';
-import { Service, ServiceOptions, HookContext, FeathersService, Application } from '../declarations';
+import {
+  getManager, HookContextData, HookManager, HookMap, HOOKS, hooks, Middleware
+} from '../dependencies';
+import {
+  Service, ServiceOptions, HookContext, FeathersService, Application
+} from '../declarations';
 import { defaultServiceArguments, getHookMethods } from '../service';
 import {
   collectLegacyHooks,